HL2 versaclock cl2 integer mode enabled
HL2 provides a clock output from the Versaclock (CL2). This can be used
to drive a transverter. This adds a new method to use this clock. The
user selects integer mode within xvtr dialog. This calculates the
nearest LO freq (from the one that was entered) which uses an integer
only divisor for the clock. This reduces spurs (may be neccesary if
using CL2 for a transverter with TX).